+++ R/get_genome.R
@@ -1,0 +1,7 @@
+get_genome <- function(build){
+    requireNamespace("regioneR")
+    regioneR::filterChromosomes(
+        A = regioneR::getGenome(build),
+        keep.chr = paste0("chr",c(seq_len(22),"X","Y")) 
+    ) 
+}