+++ R/split_chromosomes_run.R
@@ -1,0 +1,5 @@
+split_chromosomes_run <- function(query_granges){
+    GenomicRanges::split(
+        x = query_granges,
+        f = GenomicRanges::seqnames(query_granges))
+}